third_party/blink/web_tests/html5lib/generated/run-tests18-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
../resources/tests18.dat:
5

Test 5 of 36 in ../resources/tests18.dat failed. Input:
<!doctype html><html><noscript><plaintext></plaintext>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<plaintext></plaintext>"
|   <body>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <plaintext>
|       "</plaintext>"
